# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CreateTaskRequest:
    properties:
      content_md:
        type: string
      due_at:
        type: string
      priority:
        type: integer
      project_id:
        type: integer
      status:
        type: string
      title:
        maxLength: 200
        type: string
    required:
    - project_id
    - title
    type: object
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
...
  handler.TaskCreateData:
    properties:
      task:
        $ref: '#/definitions/models.Task'
    type: object
  handler.TaskCreateResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/handler.TaskCreateData'
      msg:
        type: string
    type: object
  handler.TaskDeleteData:
    properties:
      id:
        type: integer
      task_affected:
        type: integer
    type: object
  handler.TaskDeleteResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/handler.TaskDeleteData'
      msg:
        type: string
    type: object
  handler.TaskDetailResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/service.TaskDetail'
      msg:
        type: string
    type: object
  handler.TaskListData:
    properties:
      list:
        items:
          $ref: '#/definitions/service.TaskSummary'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  handler.TaskListResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/handler.TaskListData'
      msg:
        type: string
    type: object
  handler.TaskUpdateResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/models.Task'
      msg:
        type: string
    type: object
...
  handler.UpdateTaskRequest:
    properties:
      content_md:
        type: string
      priority:
        maximum: 5
        minimum: 1
        type: integer
      re_due_at:
        type: string
      re_project_id:
        type: integer
      sort_order:
        minimum: 0
        type: integer
      status:
        enum:
        - todo
        - done
        type: string
      title:
        maxLength: 200
        type: string
    type: object
...
  models.Task:
    properties:
      content_html:
        type: string
      content_md:
        type: string
      created_at:
        type: string
      due_at:
        type: string
      id:
        type: integer
      notified:
        type: boolean
      priority:
        type: integer
      project_id:
        type: integer
      sort_order:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  service.TaskDetail:
    properties:
      content_html:
        type: string
      due_at:
        type: string
      id:
        type: integer
      project_id:
        type: integer
      status:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
  service.TaskSummary:
    properties:
      due_at:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
  description: 管理API
  title: ToDoList API
  version: "1.0"
paths:
...
  /projects/{id}/tasks/{task_id}:
    get:
      consumes:
      - application/json
      description: 根据项目ID和任务ID获取任务详情
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回任务详情
          schema:
            $ref: '#/definitions/handler.TaskDetailResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: 任务不存在或项目不存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取任务详情
    patch:
      consumes:
      - application/json
      description: 更新任务的名称、内容、状态、优先级、项目和截止时间
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: 任务更新请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功，返回任务信息
          schema:
            $ref: '#/definitions/handler.TaskUpdateResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: 任务不存在或项目不存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: 任务已存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新任务
...
  /tasks:
    get:
      consumes:
      - application/json
      description: 获取指定项目下的任务列表，支持状态筛选和分页
      parameters:
      - description: 项目ID
        in: query
        name: project_id
        required: true
        type: integer
      - description: 任务状态（todo/done）
        in: query
        name: status
        type: string
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认20，最大100）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回任务列表
          schema:
            $ref: '#/definitions/handler.TaskListResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取任务列表
    post:
      consumes:
      - application/json
      description: 在指定项目下创建新任务
      parameters:
      - description: 任务创建请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回任务信息
          schema:
            $ref: '#/definitions/handler.TaskCreateResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: 任务已存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 创建任务
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定项目下的任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      - description: 项目ID
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功，返回任务ID和受影响的行数
          schema:
            $ref: '#/definitions/handler.TaskDeleteResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: 任务不存在或项目不存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 删除任务
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"